tests: make pytest put outputs in consistent location regardless of pwd
...this uses the pytest_cmdline_preparse hook to manipulate the command
line arguments to replace @rootdir@ with the actual calculated rootdir,
which seems to be the only way this is possible.

This hook is deprecated and may disappear, at which point we'll be
stuck, but until then...
